+++ front/src/components/panel/task/OptionItem.vue
@@ -108,7 +108,7 @@ import { showGlobalMessage } from "@/services/feedback/message"
 import { useInterfaceStore } from "@/stores"
 import type { InputCase } from "@/types/interfaceModel"
 import type { NullableTaskOptionValue } from "@/types/schedulerModel"
-import { makeInterfaceInputSchema } from "@/schemas/interfaceInput"
+import { makeInterfaceInputSchema } from "@/validation/interfaceInput"
 import { resolveInterfaceText } from "@/utils/interface/content"
 import { tryCatch } from "@/utils/tryCatch"
 import OptionScanSelectControl from "@/components/panel/task/OptionScanSelectControl.vue"
