+++ front/src/components/routines/SchedulerTaskRow.vue
