+++ front/src/components/routines/dialogs/SchedulerTaskDialogContentTabs.vue
