+++ front/src/components/routines/dialogs/SchedulerTaskDialogTriggerType.vue
