+++ front/src/components/settings/sections/RuntimeSettingsSection.vue
@@ -67,7 +67,7 @@
 import { computed } from "vue"
 import { useI18n } from "vue-i18n"
 import { NInputNumber, NSwitch } from "naive-ui"
-import { makeRuntimeNumberSchema } from "@/schemas/settings"
+import { makeRuntimeNumberSchema } from "@/validation/settings"
 import { useSettingsStore } from "@/stores"
 import type { SettingsModel } from "@/types/settingsModel"
