+++ front/src/components/settings/sections/UISettingsSection.vue
@@ -30,7 +30,7 @@
 import { computed } from "vue"
 import { useI18n } from "vue-i18n"
 import { NSelect } from "naive-ui"
-import { darkModeSchema, localeSchema } from "@/schemas/settings"
+import { darkModeSchema, localeSchema } from "@/validation/settings"
 import { useSettingsStore } from "@/stores"
 import type { SettingsModel } from "@/types/settingsModel"
