+++ front/src/components/settings/sections/UpdateSettingsSection.vue
@@ -73,7 +73,7 @@ import { NButton, NIcon, NInput, NSelect, NSwitch } from "naive-ui"
 import { ArrowUpCircleOutline } from "@vicons/ionicons5"
 import { checkUpdateApi, type UpdateInfo } from "@/services/api"
 import { showGlobalMessage } from "@/services/feedback/message"
-import { updateChannelSchema } from "@/schemas/settings"
+import { updateChannelSchema } from "@/validation/settings"
 import { useInterfaceStore, useSettingsStore } from "@/stores"
 import { tryCatch } from "@/utils/tryCatch"
 import type { SettingsModel } from "@/types/settingsModel"
