+++ front/src/components/tasks/StartConflictDialog.vue
