+++ front/src/validation/device.ts
