+++ front/src/validation/interfaceInput.test.ts
