+++ front/src/validation/interfaceInput.ts
