+++ front/src/validation/settings.test.ts
