+++ front/src/validation/settings.ts
