+++ front/src/views/RoutinesView.vue
@@ -44,9 +44,9 @@ import { ref, onMounted } from "vue"
 import { useI18n } from "vue-i18n"
 import { NButton, NCard, NIcon, NTabPane, NTabs } from "naive-ui"
 import { AddOutline, TimeOutline } from "@vicons/ionicons5"
-import SchedulerTaskDialog from "@/components/settings/dialogs/SchedulerTaskDialog.vue"
-import SchedulerTaskList from "@/components/settings/scheduler/SchedulerTaskList.vue"
-import SchedulerExecutionHistory from "@/components/settings/scheduler/SchedulerExecutionHistory.vue"
+import SchedulerTaskDialog from "@/components/routines/dialogs/SchedulerTaskDialog.vue"
+import SchedulerTaskList from "@/components/routines/SchedulerTaskList.vue"
+import SchedulerExecutionHistory from "@/components/routines/SchedulerExecutionHistory.vue"
 import { useSchedulerStore } from "@/stores"
 import type { ScheduledTask } from "@/types/schedulerModel"
